+++ objects/resources/tree.py
@@ -29,6 +29,8 @@ class Tree(SpriteObject):
         self.fall_speed = 2 + (random.random() * 3)
 
         super().__init__(game, x, y, tile=tile, layer=SPRITE_LAYER, image=None)
+        self.set_shadow()
+
         self.draw_rect = self.rect  # .draw_rect may be different while shaking, but .rect will stay the same
         self.fall_image = self.image
 
@@ -44,8 +46,6 @@ class Tree(SpriteObject):
 
         self.game.can_collide_list.add(self)
         self.game.can_axe_list.add(self)
-
-        self.set_shadow()
 
     def load_image(self):
         # load/set image name
@@ -179,55 +179,5 @@ class Tree(SpriteObject):
             "type":type(self).__name__,
             "topleft":(self.x, self.y),
             "image_name":self.image_name,
-            "flipped":self.flipped
-        }
-
-class ForestTree(Tree):
-    def __init__(self, game, x, y, tile, image_name=None, flipped=None):
-        super().__init__(game, x, y, tile, image_name, flipped)
-    
-    def get_spawn_weights(self):
-        return {
-            "CozyOak1":10,
-            "CozyOak2":10,
-            "CozyBirch1":20,
-            "Oak1":10,
-            "Oak2":10,
-            "Dead1":2,
-            "Dead2":2
-        }
-
-class IceTree(Tree):
-    def __init__(self, game, x, y, tile, image_name=None, flipped=None):
-        super().__init__(game, x, y, tile, image_name, flipped)
-
-    def get_spawn_weights(self):
-        return {
-            "SnowCone1":5,
-            "SnowCone2":5,
-            "SnowDead1":10,
-            "SnowDead2":10
-        }
-
-class AutumnTree(Tree):
-    def __init__(self, game, x, y, tile, image_name=None, flipped=None):
-        super().__init__(game, x, y, tile, image_name, flipped)
-
-    def get_spawn_weights(self):
-        return {
-            "Autumn1":10,
-            "Autumn2":10,
-            "CozyRough1":20,
-            "CozyAutumn1":10,
-        }
-      
-class MangroveTree(Tree):
-    def __init__(self, game, x, y, tile, image_name=None, flipped=None):
-        super().__init__(game, x, y, tile, image_name, flipped)
-
-    def get_spawn_weights(self):
-        return {
-            "CozyGreen1":10,
-            "Moss1":10,
-            "Moss2":10,
+            "flipped":self.flipped,
         }
