+++ settings.py
@@ -24,7 +24,6 @@ TILE_SIZE = 72
 TITLE = "LUMBERJACK SIMULATOR"
 BG_COLOR = LIGHT_GREY
 FPS = 60
-BIOME_NOISE_FACTOR = .0005
 
 # sprite settings
 SPRITESHEET_TILE_SIZE = 32
